fix tests and move async classes out of reader.Types
